Add Capistrano 3 support for deployment notification task
